refactor: util 패키지를 presentation 패키지 내부로 이동
