+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/calendar/CalendarScreen.kt
@@ -84,9 +84,9 @@ import org.moa.moa.presentation.ui.theme.GRAY4
 import org.moa.moa.presentation.ui.theme.MAIN
 import org.moa.moa.presentation.ui.theme.Strings
 import org.moa.moa.presentation.ui.theme.WHITE
-import org.moa.moa.util.buildMonthCells
-import org.moa.moa.util.emotionRes
-import org.moa.moa.util.formatDateTime
+import org.moa.moa.presentation.util.buildMonthCells
+import org.moa.moa.presentation.util.emotionRes
+import org.moa.moa.presentation.util.formatDateTime
 
 private object CalendarDimens {
     const val CALENDAR_FRACTION = 0.8f
