+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/record/recorder/RecorderScreen.kt
@@ -72,7 +72,7 @@ import org.moa.moa.presentation.ui.theme.Strings
 import org.moa.moa.presentation.ui.theme.WHITE
 import org.moa.moa.presentation.ui.theme.textStyle1
 import org.moa.moa.presentation.ui.theme.transparent
-import org.moa.moa.util.formatRecordTime
+import org.moa.moa.presentation.util.formatRecordTime
 
 private object RecorderDimens {
     const val RECORD_TIME_DELAY = 500L
