+++ presentation/src/iosMain/kotlin/org/moa/moa/presentation/util/formatDateTime.kt
@@ -1,4 +1,4 @@
-package org.moa.moa.util
+package org.moa.moa.presentation.util
 
 import kotlinx.datetime.LocalDateTime
 import platform.Foundation.NSCalendar
